the URL is scanned and detects a Ticker
