+++ public/MarketMood.html
@@ -1,6 +1,6 @@
-<html>
-  <body>
-    <h1> How are they feeling?</h1>
-    <script src="popup.js"></script>
-  </body>
-</html>
+<body>
+  <h1>How are they feeling?</h1>
+  <div id="ticker-result"></div>
+
+  <script src="popup.js" type="module"></script>
+</body>
